// src/app/models/travel-detail.model.ts
import { WaitDetailItemModel } from "./wait-detail-item.model";
import { ITollDetailResponse, IParkingDetailResponse, IOtherCostDetailResponse } from '../services/connection/interfaces/apiInterfaces';

export class TravelDetailModel {
  travelId: number;
  passengerFirstName: string;
  passengerLastName: string;
  passengerPhone: string;
  passengerEmail: string | null;
  hasGuestPassenger: boolean;
  guestPassengerName: string;
  requesterUser: string;
  travelDuration: number;
  driverFirstName: string;
  driverLastName: string;
  imageDriver: string;
  origin: any;
  destination: any;
  vehicleDescription: string;
  vehicleDomain: string;
  routeMap: Array<any>;
  paymentMethod: string;
  amount: number;
  hideAmountsDriver: boolean;
  tourPrice: number;
  roundTrip: boolean;
  waitTimePrice: number;
  isCorporate: boolean;
  delay: number;
  parking: number;
  pointToPointPrice: number;
  roadToll: number;
  otherCost: number;
  disputes: string;
  state: string;
  reasonForCancellation: string | null;
  kilometers: number;
  stops: number;
  note: string;
  isDelivery: boolean;
  carrierId: number;
  carrierName: string;
  carrierCode: string;
  carrierLogo: string;
  travelDate: string;
  transportType: string;
  isProgrammed: boolean;
  travelRating: number | null;
  simulationPrice: any;
  waitDetailList: Array<WaitDetailItemModel>;
  waypoints: any;
  avoidHighways: boolean;
  avoidTolls: boolean;
  fastestRoute?: boolean;
  recurringTrip: boolean;
  tollList: ITollDetailResponse[];
  parkingList: IParkingDetailResponse[];
  otherCostList: IOtherCostDetailResponse[];
  hideAmountsAffiliate: any;
  affiTotalTripCost: any;
  useDistanceMatrix?: boolean;
}